+++ ESFramework/Core/Additions/NSDictionary+ESAdditions.h
@@ -8,7 +8,7 @@
 
 #import <Foundation/Foundation.h>
 
-@interface NSDictionary (ESAdditions)
+@interface NSDictionary<KeyType, ObjectType> (ESAdditions)
 
 - (BOOL)isEmpty;
 
@@ -20,44 +20,8 @@
  */
 - (NSString *)URLQueryString;
 
-/**
- * If the object is `[NSNull null]`, it will returns `nil`.
- * It is useful when accessing a JSONObject which parsed from JSON data.
- */
-- (id)esObjectForKey:(id)key;
-
-- (void)each:(void (^)(id key, id obj, BOOL *stop))block;
-- (void)each:(void (^)(id key, id obj, BOOL *stop))block option:(NSEnumerationOptions)option;
-
-/**
- * Matches using predicating block, returns the key.
- */
-- (id)match:(BOOL (^)(id key, id obj))predicate;
-- (id)match:(BOOL (^)(id key, id obj))predicate option:(NSEnumerationOptions)option;
-
-/**
- * Returns nil if matching failed.
- */
-- (NSDictionary *)matchDictionary:(BOOL (^)(id key, id obj))predicate;
-- (NSDictionary *)matchDictionary:(BOOL (^)(id key, id obj))predicate option:(NSEnumerationOptions)option;
-
-/**
- * Matches using predicating block, returns the keys.
- */
-- (NSSet *)matches:(BOOL (^)(id key, id obj, BOOL *stop))predicate;
-- (NSSet *)matches:(BOOL (^)(id key, id obj, BOOL *stop))predicate option:(NSEnumerationOptions)option;
-
-/**
- * Matches using predicating block, returns the keys and objects.
- */
-- (NSDictionary *)matchesDictionary:(BOOL (^)(id key, id obj, BOOL *stop))predicate;
-- (NSDictionary *)matchesDictionary:(BOOL (^)(id key, id obj, BOOL *stop))predicate option:(NSEnumerationOptions)option;
-
-/**
- * Asynchronously write file.
- * It will create directories automatically if it is not exists.
- */
-- (void)writeToFile:(NSString *)path atomically:(BOOL)useAuxiliaryFile completion:(void (^)(BOOL result))completion;
+- (NSDictionary<KeyType, ObjectType> *)entriesPassingTest:(BOOL (^)(KeyType key, ObjectType obj, BOOL *stop))predicate;
+- (NSDictionary<KeyType, ObjectType> *)entriesWithOptions:(NSEnumerationOptions)opts passingTest:(BOOL (^)(KeyType key, ObjectType obj, BOOL *stop))predicate;
 
 @end
 
@@ -69,8 +33,5 @@
  * @param keyPath A key path of the form relationship.property (with one or more relationships); for example “department.name” or “department.manager.lastName”.
  */
 - (BOOL)es_setValue:(id)value forKeyPath:(NSString *)keyPath;
-
-- (void)matchWith:(BOOL (^)(id key, id obj, BOOL *stop))predicate;
-- (void)matchWith:(BOOL (^)(id key, id obj, BOOL *stop))predicate option:(NSEnumerationOptions)option;
 
 @end